rle_util_test: use int32

+++ cpp/src/arrow/util/rle_util_test.cc
@@ -27,16 +27,16 @@ namespace arrow {
 namespace rle_util {
 
 TEST(TestRleUtil, FindPhysicalOffsetTest) {
-  ASSERT_EQ(FindPhysicalOffset((const int64_t[]){1}, 1, 0), 0);
-  ASSERT_EQ(FindPhysicalOffset((const int64_t[]){1, 2, 3}, 3, 0), 0);
-  ASSERT_EQ(FindPhysicalOffset((const int64_t[]){1, 2, 3}, 3, 1), 1);
-  ASSERT_EQ(FindPhysicalOffset((const int64_t[]){1, 2, 3}, 3, 2), 2);
-  ASSERT_EQ(FindPhysicalOffset((const int64_t[]){1, 2, 3}, 3, 3), 3);
-  ASSERT_EQ(FindPhysicalOffset((const int64_t[]){2, 3, 4}, 3, 0), 0);
-  ASSERT_EQ(FindPhysicalOffset((const int64_t[]){2, 3, 4}, 3, 1), 0);
-  ASSERT_EQ(FindPhysicalOffset((const int64_t[]){2, 3, 4}, 3, 2), 1);
-  ASSERT_EQ(FindPhysicalOffset((const int64_t[]){2, 3, 4}, 3, 3), 2);
-  ASSERT_EQ(FindPhysicalOffset((const int64_t[]){2, 4, 6}, 3, 3), 1);
+  ASSERT_EQ(FindPhysicalOffset((const int32_t[]){1}, 1, 0), 0);
+  ASSERT_EQ(FindPhysicalOffset((const int32_t[]){1, 2, 3}, 3, 0), 0);
+  ASSERT_EQ(FindPhysicalOffset((const int32_t[]){1, 2, 3}, 3, 1), 1);
+  ASSERT_EQ(FindPhysicalOffset((const int32_t[]){1, 2, 3}, 3, 2), 2);
+  ASSERT_EQ(FindPhysicalOffset((const int32_t[]){1, 2, 3}, 3, 3), 3);
+  ASSERT_EQ(FindPhysicalOffset((const int32_t[]){2, 3, 4}, 3, 0), 0);
+  ASSERT_EQ(FindPhysicalOffset((const int32_t[]){2, 3, 4}, 3, 1), 0);
+  ASSERT_EQ(FindPhysicalOffset((const int32_t[]){2, 3, 4}, 3, 2), 1);
+  ASSERT_EQ(FindPhysicalOffset((const int32_t[]){2, 3, 4}, 3, 3), 2);
+  ASSERT_EQ(FindPhysicalOffset((const int32_t[]){2, 4, 6}, 3, 3), 1);
 }
 
 }  // namespace rle_util
